# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: CHAOS Engine API
  description: >
    C.H.A.O.S. -- Connected Human-Augmented OSINT Suite.
    35 sources, single binary, zero cloud.
  version: "0.1.0"
  license:
    name: AGPL-3.0-only
servers:
  - url: http://localhost:3117
    description: Local development
  - url: https://your-domain.com
    description: Production (behind Caddy reverse proxy)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-CHAOS-Key
      description: Required for extended endpoints in public mode.
...
paths:
...
  /api/v1/market-seeds:
    get:
      summary: Prediction market seeds
      description: Suggested prediction market questions generated from current data.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Market seed suggestions
          content:
            application/json:
              schema:
                type: object
...